mysite/requirements_dev.txt
-r requirements.txt


flake8==5.0.4
flake8-builtins==2.1.0
flake8-django==1.1.5
flake8-import-order==0.18.2
flake8-print==5.0.0